Move NLua code from Marble to Marble.Demo
