+++ Marble/MarbleGame.cs
@@ -1,8 +1,5 @@
-﻿using NLua;
-using NLua.Exceptions;
-using System.Collections.Immutable;
+﻿using System.Collections.Immutable;
 using System.Numerics;
-using System.Reflection;
 
 namespace Maiswan.Marble;
 
@@ -13,56 +10,12 @@ public class MarbleGame(IEnumerable<TeamBase> teams)
     public event EventHandler<MarbleGameChangedEventArgs>? GameStepped;
     public event EventHandler<MarbleGameChangedEventArgs>? GameEnded;
 
-
     private readonly List<TeamBase> teams = new(teams);
     public int TotalPopulation => teams.Sum(x => x.Population);
     public IImmutableList<TeamBase> Teams => teams.ToImmutableList();
     public IImmutableList<TeamBase> AliveTeams => teams.Where(x => x.Population != 0).ToImmutableList();
 
     private int iteration;
-
-    private readonly Lua lua = new();
-    public string LuaFile { get; init; } = "";
-
-    public void Run()
-    {
-        if (lua.IsExecuting) { return; }
-
-        RaiseGameChangedEvent(false);
-
-        // Invoke user script through sandbox
-        try
-        {
-            InitializeLuaSandbox();
-            InvokeLua();
-        }
-        catch ( LuaException ) { }
-
-        RaiseGameChangedEvent(true);
-    }
-
-    private void InitializeLuaSandbox()
-    {
-        lua["MarbleGame"] = this;
-
-        // Execute Sandbox.lua as an embdded resource
-        Assembly executing = Assembly.GetExecutingAssembly();
-        string? name = executing.GetName().Name;
-
-        using Stream? stream = executing.GetManifestResourceStream($"{name}.Sandbox.lua");
-        if (stream is null) { throw new IOException(nameof(stream)); }
-
-        StreamReader streamReader = new(stream);
-        string sandbox = streamReader.ReadToEnd();
-        lua.DoString(sandbox);
-    }
-
-    private void InvokeLua()
-    {
-        string user = File.ReadAllText(LuaFile);
-        LuaFunction sandbox = (LuaFunction)lua["run_sandbox"];
-        sandbox.Call(user);
-    }
 
     private void RaiseGameChangedEvent(bool isGameOver)
     {
@@ -79,15 +32,8 @@ public class MarbleGame(IEnumerable<TeamBase> teams)
     private void CheckEnd()
     {
         iteration++;
-
-        if (teams.Any(x => x.Population != 0))
-        {
-            RaiseGameChangedEvent(false);
-            return;
-        }
-
-        // Halt the script via an exception
-        lua.State.Error();
+        bool isGameOver = TotalPopulation == 0;
+        RaiseGameChangedEvent(isGameOver);
     }
 
     private static void ForeachTeam<T>(Func<int, T, T, int> formula, T min, T max, IEnumerable<TeamBase> teams) where T : INumber<T>
